[cern] Make cinder volume optional
Cherry-Pick: https://review.openstack.org/#/c/391830/

In the swarm_atomic and k8s_atomic drivers container images are
stored in a dedicated cinder volume per cluster node. It is
proven that this architecture can be a scalability bottleneck.

Make the use of cinder volumes for container images and opt-in
option.  If docker-volume-size is not specified no cinder
volumes will be created. Before, if docker-volume-size wasn't
specified the default value was 25.

To use cinder volumes for container storage the user will
interact with magnum as before, (meaning the valid values are
integers starting from 1).

Closes-Bug: #1638006
Change-Id: I3394c62a43bbf950b7cf0b86a71b1d9b0481d68f

Conflicts:
* magnum/drivers/common/swarm_fedora_template_def.py
  [edited magnum/drivers/swarm_fedora_atomic_v1/template_def.py instead ]
* magnum/tests/unit/conductor/handlers/test_k8s_cluster_conductor.py
  Removed the unit test for the empty docker_volume_size case.
  Fixed the new unit to not metnion the driver.